+++ History.md
@@ -5,6 +5,7 @@
 - When creating a motion, the confirmation page now shows a preview of the generated PDF.
 - Motion types can now force motion titles to have a certain beginning, like "Application: ".
 - For each section of a motion type, it is now possible to specify if the title will be explicitly printed in the PDF of not.
+- In the initiator form, the resolution date for organizations submitting a motion is now optional. An additional optional field to add one's gender was added.
 
 ### Version 4.0.1 (2018-09-02)